requirements.txt
asgiref==3.5.2
astroid==2.11.3
asttokens==2.0.5
autopep8==1.6.0
backcall==0.2.0
black==22.3.0
certifi @ file:///opt/conda/conda-bld/certifi_1655968806487/work/certifi
charset-normalizer==2.1.0
click==8.1.2
coreapi==2.3.3
coreschema==0.0.4
debugpy==1.6.0
dill==0.3.4
distlib==0.3.4
Django==3.2
django-extensions==3.2.0
django-filter==22.1
djangorestframework==3.13.1
executing==0.8.3
filelock==3.6.0
flake8==4.0.1
idna==3.3
ipykernel==6.13.0
ipython==8.2.0
isort==5.10.1
itypes==1.2.0
jedi==0.18.1
Jinja2==3.1.2
jupyter-client==7.2.2
jupyter-core==4.9.2
lazy-object-proxy==1.7.1
MarkupSafe==2.1.1
matplotlib-inline==0.1.3
mccabe==0.6.1
mypy-extensions==0.4.3
nest-asyncio==1.5.5
parso==0.8.3
pathspec==0.9.0
pickleshare==0.7.5
pipenv==2022.1.8
platformdirs==2.5.1
pure-eval==0.2.2
pycodestyle==2.8.0
pyflakes==2.4.0
Pygments==2.11.2
pylint==2.13.7
python-dateutil==2.8.2
pytz==2022.2.1
pyzmq==22.3.0
requests==2.28.1
sqlparse==0.4.2
stack-data==0.2.0
toml==0.10.2
tomli==2.0.1
tornado==6.1
typing_extensions==4.2.0
uritemplate==4.1.1
urllib3==1.26.11
virtualenv==20.13.2
virtualenv-clone==0.5.7
wrapt==1.14.0
